Added authentication

+++ BackEnd Testing/backend/controllers/stuff.js
@@ -62,19 +62,33 @@ exports.changeOneThing = (req, res, next) => {
 }
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id: req.params.id}).then(
-        () => {
-            res.status(200).json({
-                message: 'Deleted!'
-            });
+    Thing.findOne({_id: req.params.id}).then(
+      (thing) => {
+        if (!thing) {
+          return res.status(404).json({
+            error: new Error('No such thing!')
+          });
         }
-        ).catch(
-            (error) => {
-              res.status(400).json({
-                error: error
+        if (thing.userId !== req.auth.userId) {
+          return res.status(400).json({
+            error: new Error('Unauthorised request')
+          });
+        }
+        Thing.deleteOne({_id: req.params.id}).then(
+          () => {
+              res.status(200).json({
+                  message: 'Deleted!'
               });
-            }
-        );
+          }
+          ).catch(
+              (error) => {
+                res.status(400).json({
+                  error: error
+                });
+              }
+          );
+      }
+    )
 }
 
 exports.getAllStuff = (req, res, next) => {
